Add an on-hover highlight effect.
